Integrate GCF parser into package
